% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAT.R
\name{AAT}
\alias{AAT}
\title{Audio Attention Study Modules}
\usage{
AAT(
  label = "AAT",
  sub_group = "a",
  items_prefix_pattern = "s\%02d",
  num_stimuli = 1,
  random_order = FALSE,
  audio_url =
    "https://s3.eu-west-1.amazonaws.com/media.gold-msi.org/test_materials/GAR/EMO1",
  audio_type = "wav",
  allow_na = TRUE,
  dict = GAR::GAR_dict,
  ...
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}
}
\description{
This function defines module for #'
}
