% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aat_item_evaluation.R
\name{aat_item_evaluation}
\alias{aat_item_evaluation}
\title{Audio Attention Study Modules}
\usage{
aat_item_evaluation(
  label = "AAT_ITEM_EVAL",
  scale_length = 5,
  nested = F,
  dict = GAR::GAR_dict,
  max_items = 43,
  with_intro = T,
  sep = " ... ",
  ...
)
}
\arguments{
\item{label}{(character scale) Label for the whole shit}

\item{scale_length}{(integer scalar) Length of agreement scale to be used (5 or 7)}

\item{nested}{(boolean) Asked two question per item all questions per item?}

\item{dict}{(psychTestR dictionary object) You really want another dictionary?}

\item{max_items}{(integer) For debugging, only use max_items (max 43)}
}
\description{
This function defines module for aat_item_evaluation
}
