% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_radiobuttons.R
\name{audio_radiobutton_multi_NAFC_page}
\alias{audio_radiobutton_multi_NAFC_page}
\title{New radiobutton NAFC page}
\usage{
audio_radiobutton_multi_NAFC_page(
  label,
  prompt,
  items = "",
  choices,
  labels = NULL,
  trigger_button_text = "Continue",
  failed_validation_message = "Answer missing!",
  save_answer = TRUE,
  hide_response_ui = FALSE,
  response_ui_id = "response_ui",
  random_order = T,
  on_complete = NULL,
  audio_url =
    "https://s3.eu-west-1.amazonaws.com/media.gold-msi.org/test_materials/GAR/EMO1",
  audio_type = "wav",
  btn_play_prompt = if (!show_controls) "Click here to play",
  show_controls = FALSE,
  allow_download = FALSE,
  allow_na = FALSE,
  wait = TRUE,
  loop = FALSE,
  admin_ui = NULL
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{prompt}{Prompt to be displayed above the response choices.
Can be either a character scalar (e.g. "What is 2 + 2?") or an object of
class "shiny.tag", e.g. \code{shiny::tags$p("What is 2 + 2?")}.}

\item{choices}{(Character vector) Choices for the participant.
If unnamed, then these values will be used both for radiobutton IDs and for
radiobutton labels.
If named, then values will be used for radiobutton IDs and names will be used
for radiobutton labels.}

\item{labels}{Optional vector of labels for the radiobutton NAFC choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{trigger_button_text}{(Character scalar) Text for the trigger button.}

\item{failed_validation_message}{(Character scalar) Text to be displayed
when validation fails.}

\item{save_answer}{(Boolean scalar) Whether or not to save the answer.}

\item{hide_response_ui}{(Boolean scalar) Whether to begin with the response
interface hidden (it can be subsequently made visible through Javascript,
using the element ID as set in \code{response_ui_id}.
See \code{audio_NAFC_page} for an example.).}

\item{response_ui_id}{(Character scalar) HTML ID for the response user interface.}

\item{subprompt}{(Character scalar) Optional additional text in bold letters
below the prompt.}
}
\description{
Creates a radiobutton n-alternative forced choice page.
}
