\name{envGAR}
\alias{envGAR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Creates modifiable environment for package variables. 
}
\description{
Creates modifiable environment for package variables. The tokenRefresh() function will store temporary access token to GAR_ACCESS_TOKEN in the envGAR environment.

}
%\usage{
%getCode(client_id)
%}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%  \item{client_id}{
%Client ID found in API Console. String format.
%}
%}
%\details{
%Function opens a webiste URL in a browser for authenication. A code can be retrieved from the redirect URL. 
%}
%\value{
%Code is a string to be used in conjunction with gaRefresh() function.
%}
%\references{
%https://developers.google.com/analytics/devguides/reporting/core/v3/
%}
\author{
Andrew Geisler
}


