% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{GAR_standalone}
\alias{GAR_standalone}
\title{Standalone version of General Audio Rating questionnaires}
\usage{
GAR_standalone(
  label = "EMO1",
  questionnaire = "EMO1",
  response_scale = "L7",
  items_prefix_pattern = "s\%02d",
  item_order = NULL,
  num_stimuli = 10,
  num_rating_items = 6,
  anchors = FALSE,
  header = "double",
  header_style = NULL,
  reduce_labels = TRUE,
  reverse_anchors = FALSE,
  style = default_style,
  allow_na = TRUE,
  audio_url =
    "https://s3.eu-west-1.amazonaws.com/media.gold-msi.org/test_materials/GAR/EMO1",
  audio_type = "wav",
  with_id = FALSE,
  dict = GAR::GAR_dict,
  random_order = FALSE,
  randomize_stimuli = FALSE,
  show_controls = TRUE,
  allow_download = FALSE,
  admin_password = "groove",
  researcher_email = NULL,
  vertical = FALSE,
  languages = GAR::languages(),
  validate_id = "auto",
  ...
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{anchors}{(boolean)  Only for unipolar items: show scale labels?}

\item{header}{(character scale) One of simple_num, simple_str or double. simple_num only displays numbers in the header, simple_str uses verbal labels given in parameter choices, double used both.}

\item{reduce_labels}{(boolean) Shall a only the endpoints of choice labels being labelled?}

\item{reverse_anchors}{(boolean)  Reverse the scale direction}

\item{style}{(character scalar) CSS string for the radiobuttons}

\item{show_controls}{Whether or not to show audio controls to the
participant, so that they can control audio playback.}

\item{allow_download}{Whether the participant is given a button to download
the audio file; only relevant if \code{show_controls} is \code{TRUE}.}
}
\description{
This function launches a standalone testing session for the General Audio scale
}
