% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAR.R
\name{MAS_item_evaluation}
\alias{MAS_item_evaluation}
\title{MAS_item_evaluation}
\usage{
MAS_item_evaluation(
  label = "MAS_IE",
  style = default_style,
  header_style = NULL,
  random_order = FALSE,
  vertical = TRUE,
  dict = GAR::GAR_dict
)
}
\description{
This function defines a module for assessing MAS items
}
