% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{radiobutton_matrix_page}
\alias{radiobutton_matrix_page}
\title{New radio button matrix page}
\usage{
radiobutton_matrix_page(
  label,
  polarity = c("unipolar", "bipolar"),
  items,
  choices,
  instruction = NULL,
  labels = NULL,
  anchors = FALSE,
  header = c("double", "simple_str", "simple_num"),
  header_style = NULL,
  sublabel_type = c("none", "directed", "symmetric"),
  reduce_labels = TRUE,
  reverse_anchors = FALSE,
  style = default_style,
  trigger_button_text = "Continue",
  allow_na = FALSE,
  failed_validation_message = "Answer missing!",
  save_answer = TRUE,
  hide_response_ui = FALSE,
  random_order = FALSE,
  response_ui_id = "response_ui",
  on_complete = NULL,
  admin_ui = NULL
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{polarity}{("unipolar" or "bipolar") Determines whether scale will be unipolar (one label) or bipolar (two labels)}

\item{items}{(Character vector) Items to be displayed over the response
choices}

\item{choices}{(Character vector) Choices for the participant.
If unnamed, then these values will be used both for radio button IDs and for
button labels.
If named, then values will be used for button IDs and names
will be used for button labels.}

\item{labels}{Optional vector of labels for the NAFC radiobutton choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{anchors}{(boolean)  Only for unipolar items: show scale labels?}

\item{header}{(character scale) One of simple_num, simple_str or double. simple_num only displays numbers in the header, simple_str uses verbal labels given in parameter choices, double used both.}

\item{reduce_labels}{(boolean) Shall a only the endpoints of choice labels being labelled?}

\item{reverse_anchors}{(boolean)  Reverse the scale direction}

\item{style}{(character scalar) CSS string for the radiobuttons}

\item{trigger_button_text}{(Character scalar) Text for the trigger button.}

\item{failed_validation_message}{(Character scalar) Text to be displayed
when validation fails.}

\item{save_answer}{(Boolean scalar) Whether or not to save the answer.}

\item{hide_response_ui}{(Boolean scalar) Whether to begin with the response
interface hidden (it can be subsequently made visible through Javascript,
using the element ID as set in \code{response_ui_id}.
See \link[psychTestR]{audio_NAFC_page} for an example.).}

\item{response_ui_id}{(Character scalar) HTML ID for the response user interface.}

\item{on_complete}{Optional function to execute on leaving the page
(after successful validation).
The argument list should include \code{...},
and any of:
\code{state}, the participant's state object;
\code{answer}, the participant's most recent answer;
\code{input}, the current page's Shiny input object;
\code{session}, the current Shiny session object;
\code{opt}, the test's option list as created by \code{test_options()}.}

\item{admin_ui}{Optional UI component for the admin panel.}

\item{prompts}{(Character vector) Prompts to be displayed over the response
choices}
}
\description{
Creates a  radio button matrix pag for n items with m Likert-type choices
}
