% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAT.R
\name{AAT}
\alias{AAT}
\title{Audio Attention Study Modules}
\usage{
AAT(
  label = "AAT",
  sub_group = "a",
  items_prefix_pattern = "s\%03d",
  num_stimuli = 1,
  random_order = FALSE,
  with_module = F,
  header_style = NULL,
  audio_url = "https://s3.eu-west-1.amazonaws.com/media.dots.org/stimuli/AAT",
  audio_type = "wav",
  allow_na = TRUE,
  dict = GAR::GAR_dict,
  ...
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{sub_group}{(character scalar, "a", "b", "c" or "d")
One of four subgroups (bipolar/unipolar, switched labels)}

\item{num_stimuli}{(integer)
Number of stimuli}

\item{random_order}{(boolean)
Randomize items (not stimuli!)}

\item{with_module}{(boolean)
Wrap AAt into module.}

\item{header_style}{(string)
Header style to be pass on to make_ui_radiobutton_NAFC}

\item{audio_url}{(URL)
Where are the stimuli?}

\item{audio_type}{(character scalar)
Audio type of stimuli (as file extension)}

\item{allow_na}{(boolean)
Currently ignored}

\item{dict}{(psychTestR dictionary object)
You really want another dictionary?}

\item{item_prefix_pattern}{(format string)
Used for identifying items using audio_url and audio_type}
}
\description{
This function defines module for #'
}
