% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{audio_radiobutton_matrix_page}
\alias{audio_radiobutton_matrix_page}
\title{Make radiobutton matrix audio page}
\usage{
audio_radiobutton_matrix_page(
  label,
  polarity = c("unipolar", "bipolar"),
  prompts,
  choices,
  url,
  instruction = "",
  labels = NULL,
  anchors = FALSE,
  header = "double",
  style = list(label_widths = list("100px", "100px"), div_style =
    "width:60\%;margin-left:20\%;margin-right:20\%"),
  reduce_labels = TRUE,
  trigger_button_text = "Continue",
  allow_na = FALSE,
  failed_validation_message = "Answer missing!",
  save_answer = TRUE,
  hide_response_ui = FALSE,
  random_order = FALSE,
  response_ui_id = "response_ui",
  on_complete = NULL,
  audio_type = tools::file_ext(url),
  wait = TRUE,
  loop = FALSE,
  admin_ui = NULL,
  btn_play_prompt = if (!show_controls) "Click here to play",
  show_controls = FALSE,
  allow_download = FALSE
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{prompts}{(Character vector) Prompts to be displayed over the response
choices}

\item{choices}{(Character vector) Choices for the participant.
If unnamed, then these values will be used both for radio button IDs and for
button labels.
If named, then values will be used for button IDs and names
will be used for button labels.}

\item{url}{URL to the audio. Can be an absolute URL (e.g.
"http://mysite.com/audio.mp3") or a URL relative to the /www directory (e.g.
"audio.mp3").}

\item{labels}{Optional vector of labels for the NAFC radiobutton choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{trigger_button_text}{(Character scalar) Text for the trigger button.}

\item{failed_validation_message}{(Character scalar) Text to be displayed
when validation fails.}

\item{save_answer}{(Boolean scalar) Whether or not to save the answer.}

\item{hide_response_ui}{(Boolean scalar) Whether to begin with the response
interface hidden (it can be subsequently made visible through Javascript,
using the element ID as set in \code{response_ui_id}.
See \link[psychTestR]{audio_NAFC_page} for an example.).}

\item{response_ui_id}{(Character scalar) HTML ID for the response user interface.}

\item{on_complete}{Optional function to execute on leaving the page
(after successful validation).
The argument list should include \code{...},
and any of:
\code{state}, the participant's state object;
\code{answer}, the participant's most recent answer;
\code{input}, the current page's Shiny input object;
\code{session}, the current Shiny session object;
\code{opt}, the test's option list as created by \code{test_options()}.}

\item{wait}{Whether to wait for the audio to finish before displaying the
response buttons.}

\item{loop}{Whether the audio should loop.}

\item{admin_ui}{Optional UI component for the admin panel.}

\item{show_controls}{Whether or not to show audio controls to the
participant, so that they can control audio playback.}

\item{allow_download}{Whether the participant is given a button to download
the audio file; only relevant if \code{show_controls} is \code{TRUE}.}

\item{type}{Audio type (e.g. 'mp3'). Defaults to the provided file extension.}
}
\description{
Creates a  radiobutton matrix page for a single audio file.
}
