% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAR.R, R/general_audio_ratings.R
\docType{package}
\name{GAR}
\alias{GAR}
\title{General Audio Ration component}
\usage{
GAR(
  label = "EMO1",
  type = "EMO1",
  response_scale = "L7",
  items_prefix_pattern = "s\%02d",
  num_stimuli = 10,
  num_rating_items = 6,
 
    audio_url = "https://s3.eu-west-1.amazonaws.com/media.gold-msi.org/test_materials/GAR/EMO1",
  audio_type = "wav",
  dict = GAR::GAR_dict,
  random_order = FALSE,
  show_controls = TRUE,
  allow_download = FALSE,
  ...
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{type}{Audio type (e.g. 'mp3'). Defaults to the provided file extension.}

\item{random_order}{(Boolean scalar) Whether the order of the items should
be randomized.}

\item{show_controls}{Whether or not to show audio controls to the
participant, so that they can control audio playback.}

\item{allow_download}{Whether the participant is given a button to download
the audio file; only relevant if \code{show_controls} is \code{TRUE}.}
}
\description{
This function defines a module for rating music or audio excerpts on a list of Likert items;
for incorporation into a
psychTestR timeline. Use this function if you want to include GAR in a
battery of other tests, or if you want to add custom psychTestR pages to
your test timeline. For a standalone implementation of the GAR, consider
using \code{GAR_standalone()}.

Contains the several simple audio rating scales.
}
\author{
Klaus Frieler \email{klaus.frieler@ae.mpg.de}
}
