% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{GAR_standalone}
\alias{GAR_standalone}
\title{Standalone version of General Audio Rating questionnaires}
\usage{
GAR_standalone(
  label = "EMO1",
  type = "EMO1",
  response_scale = "L7",
  items_prefix_pattern = "s\%02d",
  num_stimuli = 10,
  num_rating_items = 6,
 
    audio_url = "https://s3.eu-west-1.amazonaws.com/media.gold-msi.org/test_materials/GAR/EMO1",
  audio_type = "wav",
  with_id = FALSE,
  dict = GAR::GAR_dict,
  random_order = FALSE,
  show_controls = TRUE,
  allow_download = FALSE,
  admin_password = "groove",
  researcher_email = NULL,
  languages = GAR::languages(),
  validate_id = "auto",
  ...
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{type}{Audio type (e.g. 'mp3'). Defaults to the provided file extension.}

\item{random_order}{(Boolean scalar) Whether the order of the items should
be randomized.}

\item{show_controls}{Whether or not to show audio controls to the
participant, so that they can control audio playback.}

\item{allow_download}{Whether the participant is given a button to download
the audio file; only relevant if \code{show_controls} is \code{TRUE}.}
}
\description{
This function launches a standalone testing session for the General Audio scale
}
