% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_elements.R
\name{make_ui_radiobutton_matrix}
\alias{make_ui_radiobutton_matrix}
\title{New radio button matrix page}
\usage{
make_ui_radiobutton_matrix(
  label,
  polarity = c("unipolar", "bipolar"),
  items = NULL,
  scale_labels = NULL,
  choices = 1:length(scale_labels),
  reduce_labels = TRUE,
  anchors = TRUE,
  style = default_style,
  header = c("simple_num", "simple_str", "double"),
  trigger_button_text = "Continue",
  hide = FALSE,
  id = "response_ui"
)
}
\arguments{
\item{label}{(Character scalar) Label for the current page.}

\item{items}{(Character vector) Items to be displayed over the response
choices}

\item{scale_labels}{Optional vector of labels for the radio button choices.
If not \code{NULL}, will overwrite the names of \code{choices}.
This vector of labels can either be a character vector
or a list of Shiny tag objects, e.g. as created by \code{shiny::HTML()}.}

\item{choices}{(Character vector) Choices for the participant.
If unnamed, then these values will be used both for radio button IDs and for
button labels.
If named, then values will be used for button IDs and names
will be used for button labels.}

\item{trigger_button_text}{(Character scalar) Text for the trigger button.}

\item{hide}{(Boolean scalar) Whether to begin with the response
interface hidden (it can be subsequently made visible through Javascript,
using the element ID as set in \code{response_ui_id}.
See \link[psychTestR]{audio_NAFC_page} for an example.).}

\item{id}{(character scalar) ID for the response ui}
}
\description{
Creates a  radio button matrix pag for n items with m Likert-type choices
}
